# docs/swagger.yaml
definitions:
...
  controllers.ReviewInput:
    properties:
      game_id:
        type: integer
      text:
        type: string
      user_id:
        type: integer
    type: object
  controllers.TopReviewInput:
    properties:
      name:
        type: string
      review_ID:
        type: integer
    type: object
...
  models.Game:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      year:
        type: integer
    type: object
...
  models.Review:
    properties:
      created_at:
        type: string
      game_id:
        type: integer
      id:
        type: integer
      text:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.TopReview:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      review_ID:
        type: integer
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /Review/{id}:
    delete:
      description: Delete a Review by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Review.
      tags:
      - Review
  /Reviews:
    get:
      description: Get a list of Reviews.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
      summary: Get all Reviews.
      tags:
      - Review
    post:
      description: Creating a new Review.
      parameters:
      - description: the body to create a new Review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.ReviewInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - BearerToken: []
      summary: Create New Review.
      tags:
      - Review
  /Reviews/{id}:
    get:
      description: Get a Review by id.
      parameters:
      - description: Review id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      summary: Get Review.
      tags:
      - Review
    patch:
      description: Update Review by id.
      parameters:
      - description: Review id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an Review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.ReviewInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - BearerToken: []
      summary: Update Review.
      tags:
      - Review
  /Reviews/{id}/Games:
    get:
      description: Get Review by Game.
      parameters:
      - description: game id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Game'
            type: array
      summary: Get Review By Games.
      tags:
      - Review
  /Reviews/{id}/Users:
    get:
      description: Get Review by User.
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Get Review By User.
      tags:
      - Review
  /TopReview:
    get:
      description: Get a list of TopReview.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TopReview'
            type: array
      summary: Get all TopReview.
      tags:
      - TopReview
    post:
      description: Creating a new TopReview.
      parameters:
      - description: the body to create a new TopReview
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.TopReviewInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TopReview'
      security:
      - BearerToken: []
      summary: Create New TopReview.
      tags:
      - TopReview
  /TopReview/{id}:
    delete:
      description: Delete a TopReview by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: TopReview id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one TopReview.
      tags:
      - TopReview
    patch:
      description: Update TopReview by id.
      parameters:
      - description: TopReview id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update age TopR TopReview
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.TopReviewInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TopReview'
      security:
      - BearerToken: []
      summary: Update TopReview.
      tags:
      - TopReview
...
swagger: "2.0"